requirements.txt
django==2.1.2
djangorestframework==3.8.2
gunicorn==19.9.0
requests==2.20.1
